fix the kdtree rotate dimension

+++ tests/test_framework.h
@@ -24,7 +24,7 @@
 #ifdef CCP
 using Coord = long;
 #else
-using Coord = double;
+using Coord = long;
 #endif  // CCP
 
 using Typename = Coord;
@@ -1066,9 +1066,11 @@ void PrintTreeParam() {
             << "Inba: " << TreeWrapper::TreeType::GetImbalanceRatio() << "; ";
 
   if constexpr (std::is_integral_v<typename TreeWrapper::Point::Coord>) {
-    std::cout << "Coord: integer" << "; ";
+    std::cout << "Coord: integer"
+              << "; ";
   } else if (std::is_floating_point_v<typename TreeWrapper::Point::Coord>) {
-    std::cout << "Coord: float" << "; ";
+    std::cout << "Coord: float"
+              << "; ";
   }
   std::cout << "\n" << std::flush;
   return;
